Issue #635: move parser file
