+++ deepreg/config/__init__.py
@@ -1,0 +1,1 @@
+"""Module for configuration files."""
